Minor reorganization and get the system working where all data is stored in IPFS filesystem-style
