+++ config/config.exs
@@ -1,7 +1,7 @@
 use Mix.Config
 
 config :daisy,
-  serializer: Daisy.Serializer.JSONSerializer,
+  serializer: Daisy.Serializer,
   run_leader: false,
   run_follower: false,
   run_api: false,
