--- test/daisy/serializer/json_serializer_test.exs
@@ -1,5 +1,0 @@
-defmodule Daisy.Serializer.JSONSerializerTest do
-  use ExUnit.Case, async: true
-  doctest Daisy.Serializer.JSONSerializer
-
-end
