+++ test/daisy/serializer_test.exs
@@ -1,0 +1,5 @@
+defmodule Daisy.SerializerTest do
+  use ExUnit.Case, async: true
+  doctest Daisy.Serializer
+
+end
